update: rewritten adapters(old infra) to correspond with new database tables design
