+++ domain/model.py
@@ -1,9 +1,24 @@
+from dataclasses import dataclass
+from datetime import datetime
+
+
+@dataclass
 class AuthUser:
     def __init__(self, user_id, username, password, superuser=False):
         self.id = user_id
         self.username = username
         self.password = password
         self.superuser = superuser
+    id: int
+    username: str
+    password: str
+
+
+@dataclass(frozen=True)
+class Role:
+    id: int
+    role: str
+
 
     def __repr__(self):
         return f"<AuthUser {self.id} {self.username}>"
